% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcmp-package.R
\docType{data}
\name{attendance}
\alias{attendance}
\title{Attendance data set}
\format{
A data frame with 314 observations on 5 variables.
\describe{
\item{id}{Identifier}
\item{gender}{gender}
\item{math}{standardized math score out of 100}
\item{daysabs}{number of days absent from high school}
\item{prog}{academic programme ('General', 'Academic' and 'Vocational')}
}
}
\source{
\url{https://stats.oarc.ucla.edu/stat/stata/dae/nb_data.dta}
}
\usage{
data(attendance)
}
\value{
A data frame with 314 rows and 5 variables.
}
\description{
This data set gives the number of days absent
from high school and the gender, maths score (standardized score out of 100) and
academic programme ('General', 'Academic' and 'Vocational') of 314 students
sampled from two urban high schools. The attendance data frame has 314 observations
on 5 variables.
}
\examples{
## For examples see example(glm.cmp)
}
\keyword{datasets}
