% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{CBIND}
\alias{CBIND}
\title{Combine R Objects by Columns}
\usage{
CBIND(..., deparse.level = 1)
}
\arguments{
\item{...}{(generalized) vectors or matrices. These can be given as named arguments}

\item{deparse.level}{integer; deparse.level = 0 constructs no labels,
deparse.level = 1 (the default) or > 1 constructs labels from the arguments names.}
}
\value{
A matrix formed by column-binding the (recycled) arguments; see
\code{\link{cbind}}.
}
\description{
Take a sequence of vector, matrix or data-frame arguments and combine them by columns.
\code{CBIND} is used within the package over \code{cbind} to recycle the shorter
arguments so that their number of rows would match.
}
\keyword{internal}
