% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{coef.cmp}
\alias{coef.cmp}
\title{Extract Model Coefficients from a COM-Poisson Model Fit}
\usage{
\method{coef}{cmp}(object, ...)
}
\arguments{
\item{object}{an object class 'cmp' object, obtained from a call to \code{glm.cmp}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
Coefficients extracted from the object \code{object}.
}
\description{
An function used to extract model coefficients from a 'cmp' object.
\code{coefficients} is an alias for \code{coef}.
}
\seealso{
\code{\link{fitted.cmp}}, \code{\link{residuals.cmp}}, \code{\link{glm.cmp}}.
}
