% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_related.R
\name{compPIT}
\alias{compPIT}
\title{Non-randomized Probability Integral Transform}
\usage{
compPIT(object, bins = 10)
}
\arguments{
\item{object}{an object class "cmp", obtained from a call to \code{glm.cmp}.}

\item{bins}{numeric: the number of bins used in the PIT.}
}
\value{
compPredProb returns a list with values:
\item{upper}{the predictive cumulative probabilities used as the upper bound for computing the non-randomized PIT.}
\item{lower}{the predictive cumulative probabilities used as the lower bound for computing the non-randomized PIT.}

compPIT returns a list with values:
\item{upper}{the predictive cumulative probabilities used as the upper bound for 
computing the non-randomized PIT.}
\item{lower}{the predictive cumulative probabilities used as the lower bound for 
computing the non-randomized PIT.}
\item{conditionalPIT}{the conditional probability integral transformation 
given the observed counts.}
\item{PIT}{the probability integral transformation.}
}
\description{
Functions to produce the non-randomized probability integral transform (PIT) to check the
adequacy of the CMP distribution.
}
\details{
These functions are used for the assessment of predictive distributions 
in discrete data. They obtain the predictive probabilities and the probability 
integral transformation for a fitted CMP model.
}
\examples{

}
\references{
Czado, C., Gneiting, T. and Held, L. (2009) Predictive model assessment
for count data. \emph{Biometrics}, \strong{65}, 1254--1261.
Jung, R.C and Tremayne, A.R (2011) Useful models for time series of counts or simply wrong ones? \emph{Advances in Statistical Analysis}, \strong{95}, 59–-91.
}
