% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{fitted.cmp}
\alias{fitted.cmp}
\title{Extract Fitted Values from a COM-Poisson Model Fit}
\usage{
\method{fitted}{cmp}(object, ...)
}
\arguments{
\item{object}{an object class 'cmp' object, obtained from a call to \code{glm.cmp}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
Fitted values \code{mu} extracted from the object \code{object}.
}
\description{
An accessor function used to extract the fitted values from a 'cmp' object.
\code{fitted.values} is an alias for \code{fitted}.
}
\seealso{
\code{\link{coef.cmp}}, \code{\link{residuals.cmp}}, \code{\link{glm.cmp}}.
}
