% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{format_perc}
\alias{format_perc}
\title{Imported from the stats package}
\usage{
format_perc(
  x,
  digits = max(2L, getOption("digits")),
  probability = TRUE,
  use.fC = length(x) < 100,
  ...
)
}
\description{
Imported from the stats package
}
\keyword{internal}
