% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.cmp.R
\name{glm.cmp}
\alias{glm.cmp}
\title{Fit a Mean Parametrized Conway-Maxwell Poisson Generalized Linear Model}
\usage{
glm.cmp(formula, data, offset = NULL,
   lambdalb = 1e-10, lambdaub = 1299, maxlambdaiter = 1e3, tol = 1e-6)
}
\arguments{
\item{formula}{an object of class 'formula': a symblic desciption of the model to be 
fitted.}

\item{data}{an optional data frame containing the variables in the model}

\item{offset}{this can be used to specify an *a priori* known component to be included 
in the linear predictor during fitting. This should be \code{NULL} or a numeric vector 
of length equal to the number of cases.}

\item{lambdalb, lambdaub}{numeric: the lower and upper end points for the interval to be
searched for lambda(s).}

\item{maxlambdaiter}{numeric: the maximum number of iterations allowed to solve 
for lambda(s).}

\item{tol}{numeric: the convergence threshold. A lambda is said to satisfy the 
mean constraint if the absolute difference between the calculated mean and a fitted
values is less than tol.}
}
\value{
A fitted model object of class \code{cmp} similar to one obtained from \code{glm} 
or \code{glm.nb}.

The function \code{summary} (i.e., \code{\link{summary.cmp}}) can be used to obtain 
and print a summary of the results. 

The function \code{plot} (i.e., \code{\link{plot.cmp}}) can be used to produce a range 
of diagnostic plots. 

The generic assessor functions \code{coefficients} (i.e., \code{\link{coefficients.cmp}}), 
\code{logLik} (i.e., \code{\link{logLik.cmp}}) 
\code{fitted.values} (i.e., \code{\link{fitted.values.cmp}}), 
\code{nobs} (i.e., \code{\link{nobs.cmp}}), 
\code{AIC} (i.e., \code{\link{AIC.cmp}}) and 
\code{residuals} (i.e., \code{\link{residuals.cmp}}) 
can be used to extract various useful features of the value
returned by \code{glm.cmp}.

An object class 'glm.cmp' is a list containing at least the following components:

\item{coefficients}{a named vector of coefficients}
\item{residuals}{the *response* residuals (i.e., observed-fitted)}
\item{fitted values}{the fitted mean values}
\item{rank}{the numeric rank of the fitted linear model}
\item{linear.predictors}{the linear fit on log scale}
\item{df.residuals}{the residuals degrees of freedom}
\item{df.null}{the residual degrees of freedom for the null model}
\item{null.deviance}{The deviance for the null model. 
The null model will include only the intercept.}
\item{y}{the \code{y} vector used.}
\item{x}{the model matrix}
\item{model}{the model frame}
\item{call}{the matched call}
\item{formula}{the formula supplied}
\item{terms}{the \code{\link{terms}} object used}
\item{data}{the \code{data} argument}
\item{offset}{the \code{offset} vector used}
}
\description{
The function \code{glm.cmp} is used to fit a mean parametrized Conway-Maxwell Poisson
generalized linear model with a log-link by using Fisher Scoring iteration.
}
\examples{
### Huang (2017) Page 368--370: Overdispersed Attendance data
data(attendance)
M.attendance = glm.cmp(y~., data=attendance)
M.attehdance
summary(M.attendance)

### Huang (2017) Page 371--372: Underdispersed Takeover Bids data
data(takeover)
M.takeover = glm.cmp(numbids~., data=takeover)
M.takeover
summary(M.takeover)
}
\references{
Fung, T., Alwan, A., Wishart, J. and Huang, A. (2018). The \code{mpcmp} package for 
Mean-parametrized Conway-Maxwell Poisson Regression. 

Huang, A. (2017). Mean-parametrized Conway–Maxwell–Poisson regression models for 
dispersed counts. \emph{Statistical Modelling} \bold{17}, 359--380.
}
\seealso{
\code{\link{summary.cmp}}, \code{\link{plot.cmp}}, \code{\link{fitted.values.cmp}} 
and \code{\link{residuals.cmp}}.
}
