% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{has_rownames}
\alias{has_rownames}
\title{Imported from the broom package.
Notice that this is difference to the same function in tibble.}
\usage{
has_rownames(df)
}
\description{
Imported from the broom package.
Notice that this is difference to the same function in tibble.
}
\keyword{internal}
