% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{logLik.cmp}
\alias{logLik.cmp}
\alias{print.logLik.cmp}
\title{Extract the (Maximized) Log-Likelihood from a COM-Poisson Model Fit}
\usage{
\method{logLik}{cmp}(object, ...)

\method{print}{logLik.cmp}(x, ...)
}
\arguments{
\item{object}{an object of class 'cmp' object, obtained from a call to \code{glm.cmp}}

\item{...}{other arguments passed to or from other methods  (currently unused).}

\item{x}{an object of class 'logLik.cmp', obtained from a call to \code{logLik.cmp}.}
}
\value{
\code{logLik.cmp} returns an object of class \code{"logLik.cmp"}: the
maximized log-likelihood value with a \code{"df"} attribute giving the
number of estimated parameters, analogous to \code{\link[stats]{logLik}}.

\code{print.logLik.cmp} is called for its side effect of printing and
returns \code{x} invisibly.
}
\description{
An accessor function used to extract the (maximized) log-likelihood from a 'cmp' object.
}
\seealso{
\code{\link{coef.cmp}}, \code{\link{fitted.cmp}}, \code{\link{glm.cmp}}
}
