% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{print.cmp}
\alias{print.cmp}
\title{Print Values of COM-Poisson Model}
\usage{
\method{print}{cmp}(x, ...)
}
\arguments{
\item{x}{an object class 'cmp', obtained from a call to \code{glm.cmp}.}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
\code{x} is returned invisibly; \code{print.cmp} is called for its
side effect of printing the model summary to the console.
}
\description{
\code{print} method for class \code{cmp}.
}
\details{
\code{print.cmp} can be used to print a short summary of object class 'cmp'.
}
\examples{
## For examples see example(glm.cmp)
}
\seealso{
\code{\link{summary.cmp}}, \code{\link{coef.cmp}}, \code{\link{fitted.cmp}}, \code{\link{glm.cmp}}.
}
