% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_related.R
\name{print.cmp_autoplot_list}
\alias{print.cmp_autoplot_list}
\title{Print Method for a \code{cmp_autoplot_list} Object}
\usage{
\method{print}{cmp_autoplot_list}(x, ...)
}
\arguments{
\item{x}{an object of class \code{cmp_autoplot_list}, as returned by
\code{\link{autoplot.cmp}} with \code{output_as_ggplot = TRUE}.}

\item{...}{other arguments passed to or from other methods (currently unused).}
}
\value{
\code{x} is returned invisibly; \code{print.cmp_autoplot_list} is
called for its side effect of drawing the combined diagnostic plot on the
current graphics device.
}
\description{
Re-displays the combined diagnostic plot produced by \code{\link{autoplot.cmp}}
when \code{output_as_ggplot = TRUE}. This avoids the default list-print
behaviour (\code{[[1]]}, \code{[[2]]}, ... headers with each plot printed
separately) when the list of \code{ggplot} objects returned by
\code{autoplot.cmp}/\code{gg_plot} is printed again, e.g. after being
assigned to a variable.
}
