% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{vcov.cmp}
\alias{vcov.cmp}
\title{Extracting the Variance-Covariance Matrix from a COM-Poisson
Model Fit}
\usage{
\method{vcov}{cmp}(object, ...)
}
\arguments{
\item{object}{an object class 'cmp' object, obtained from a call to \code{glm.cmp}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
The method will return the estimated covariances between the parameter estimates of the fitted cmp model.
}
\description{
Extracting the Variance-Covariance Matrix from a COM-Poisson
Model Fit
}
\examples{
data(attendance)
M.attendance <- glm.cmp(daysabs~ gender+math+prog, data=attendance)
vcov(M.attendance)

}
