% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z.R
\name{Z}
\alias{Z}
\title{Calculate the Normalizing Constant for COM-Poisson distribution}
\usage{
Z(lambda, nu, log.z = FALSE, summax)
}
\arguments{
\item{lambda}{rate parameter, straightly positive}

\item{nu}{dispersion parameter, straightly positive}

\item{log.z}{logical; if \code{TRUE}, normalising constant \eqn{Z} are returned as
\eqn{log(Z)}.}

\item{summax}{maximum number of terms to be considered in the truncated sum}
}
\value{
A numeric vector giving the (approximate) COM-Poisson normalizing
constant \eqn{Z(\lambda, \nu)} (or its natural log if \code{log.z = TRUE}),
recycled to the length of the longer of \code{lambda} and \code{nu}.
}
\description{
A function to approximate the normalizing constant for COM-Poisson distributions
via truncation. The standard COM-Poisson parametrization is being used here.
Notice that the sum is hard coded to truncate at 100 so the approximation will be quite
bad if the COM-Poisson has a large rate or mean.
}
