% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_mu_loglik.R
\name{comp_mu_loglik}
\alias{comp_mu_loglik}
\title{Calculate the Log-Likelihood of the COM-Poisson model}
\usage{
comp_mu_loglik(param, y, xx, offset)
}
\arguments{
\item{param}{numeric vector: the model coefficients & the current value of \code{nu}. 
It is assumed that \code{nu} is in the last position of \code{param}.}

\item{y}{numeric vector: response variable}

\item{xx}{numeric matrix: the explanatory variables}

\item{offset}{numeric vector: a vector of lenght equal to the number of cases}
}
\value{
The log-likelihood value of the COM-Poisson model.
}
\description{
A function to compute the log-likelihood of the COM-Poisson model.
}
