% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Test for a whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric vector to be tested}

\item{tol}{numeric; precision level}
}
\value{
A logical vector, the same length as \code{x}, indicating whether
each element is (within tolerance \code{tol} of) a whole number.
}
\description{
Test for integer/whole number vector
}
\keyword{internal}
