% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcmp-package.R
\docType{package}
\name{mpcmp-package}
\alias{mpcmp-package}
\alias{mpcmp}
\title{Mean-parametrized Conway-Maxwell Poisson Regression}
\description{
Mean-parametrized Conway-Maxwell Poisson Regression
}
\references{
Fung, T., Alwan, A., Wishart, J. and Huang, A. (2019). \code{mpcmp}:
Mean-parametrized Conway-Maxwell Poisson Regression. R package version 0.2.0.

Huang, A. (2017). Mean-parametrized Conway-Maxwell-Poisson regression models for
dispersed counts. \emph{Statistical Modelling} \bold{17}, 359--380.
}
\keyword{package}
