% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_extract.R
\name{nobs.cmp}
\alias{nobs.cmp}
\title{Extract the Number of Observation from a COM-Poisson Model Fit}
\usage{
\method{nobs}{cmp}(object, ...)
}
\arguments{
\item{object}{an object class 'cmp' object, obtained from a call to \code{glm.cmp}}

\item{...}{other arguments passed to or from other methods  (currently unused).}
}
\value{
The number of observations extracted from the object \code{object}.
}
\description{
An accessor function used to extract the number of observation from a 'cmp' object.
}
\seealso{
\code{\link{coef.cmp}}, \code{\link{fitted.cmp}}, \code{\link{glm.cmp}}
}
