% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcmp-package.R
\docType{data}
\name{takeoverbids}
\alias{takeoverbids}
\title{Takeover Bids data set}
\format{A data frame with 126 observations on 14 variables.
\describe{
\item{bidprem}{bid price divided by price 14 working days before bid}
\item{docno}{doc no.}
\item{finrest}{indicator variable for proposed change in ownership structure}
\item{insthold}{percentage of stock held by institutions}
\item{leglrest}{indicator variable for legal defence by lawsuit}
\item{numbids}{number of bids received}
\item{obs}{Identifier}
\item{rearest}{indicator variable for proposed changes in asset structure}
\item{regulartn}{indicator variable for Department of Justice intervention}
\item{size}{total book value of assets in billions of dollars}
\item{takeover}{Indicator. 1 if the company was being taken over}
\item{weeks}{}
\item{whtknght}{indicator varible for management invitation 
for friendly third-party bid}
}}
\usage{
data(wallabies)
}
\description{
This data set gives gives the number of bids received by 126 US firms that were successful
targets of tender offers during the period 1978--1985, along with some explanatory 
variables on the defensive actions taken by management of target firm, firm-specific
characteristics and any intervention taken by federal regulators. The \code{takeoverbids} 
data frame has 126 observations on 14 variables.
}
\details{
Macropods defaecate randomly as they forage and scat 
  (faecal pellet) surveys are a reliable method for detecting the
  presence of rock-wallabies and other macropods. 
  Scats are used as an indication of spatial foraging patterns 
  of rock-wallabies and sympatric macropods. Scats deposited while
  foraging were not confused with scats deposited while
  resting because the daytime refuge areas of rock-wallabies
  were known in detail for each colony and no samples were
  taken from those areas. Each of the 200 sites were 
  examined separately to
  account for the different levels of predation risk and the
  abundance of rock-wallabies.
}
\examples{
data(wallabies)
wdat = data.frame(subset(wallabies,select=-c(lat,long)), 
  EaD = wallabies$edible*wallabies$distance,
  EaS = wallabies$edible*wallabies$shelter,
  DaS = wallabies$distance*wallabies$shelter)
M1 = glm(rw~., family = binomial(link = "logit"), data = wdat)
}
\references{
Tuft KD, Crowther MS, Connell K, Mueller S and McArthur C (2011), 
   Predation risk and competitive interactions affect foraging of 
   an endangered refuge-dependent herbivore. Animal Conservation, 
   14: 447-457. doi: 10.1111/j.1469-1795.2011.00446.x
}
\keyword{datasets}
