% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z.R
\name{Z}
\alias{Z}
\title{Calculate the Normalizing Constant for COM-Poisson distribution}
\usage{
Z(lambda, nu)
}
\arguments{
\item{lambda}{rate parameter, straightly positive}

\item{nu}{diepsersoin parameter, straightly positive}
}
\description{
A function to approximate the normalizing constant for COM-Poisson distributions
via truncation. The standard COM-Poisson parametrization is being used here. 
Notice that the sum is hard coded to tuncate at 100 so the approximation will be quite
bad if the COM-Poisson has a large rate or mean.
}
